Added build script for use on FreeBSD.
This script is mostly based on the Makefile from the FreeBSD
audio/logitechmediaserver port, but with some tweaks.

This does not currently install the files onto your system; it prints
out instructions at the end for final steps to 'sudo chown ...', copy
things to the right places, and soforth.

It also makes some assumptions about the locations of files. You should
have the 'vendor' repo available, since it is needed for the build. If
the script can't find things, it will tell you and error.